Refactor CustomPrefabUtility
