+++ Editor/Scripts/CustomPrefabUtility.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 69b7e9aef2aaf07418e068c14f4ed376
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
