+++ Editor/Scripts/CustomPrefabUtility/PrefabManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fe2a278983383f44a8750c6ddc463983
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
