+++ Editor/Scripts/CustomPrefabUtility/PrefabPreview.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f6f9eee588a3ab04888a1d9cf2b74aa2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
