+++ Editor/Scripts/CustomPrefabUtility/RenamePrefabWindow.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 750b8b809d17ef449a59fef1dcd9d761
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
