+++ Editor/Scripts/CustomPrefabUtility/ShaderChecker.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4130e7d24ccb8fd4ea20064ae5aa477c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
