+++ Editor/Scripts/CustomPrefabUtility/Utility.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 19e8f47c2cd44c047b335e5ebb486cd2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
